backend/requirements.txt
Flask
python-binance
Flask-Cors
pandas
